refactor(app-memory): split platform sampling modules
